requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dash~=3.0.4
et_xmlfile==2.0.0
Flask==3.0.3
fonttools==4.57.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.33.0
nest-asyncio==1.6.0
numpy~=2.2.5
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow~=11.2.1
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
retrying==1.3.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
psutil~=7.0.0
fpdf~=1.7.2
jsonschema~=4.23.0
scapy~=2.6.1
imbalanced-learn~=0.13.0
transformers~=4.51.3
networkx~=3.3
torch~=2.7.0+cpu
torchvision~=0.22.0+cpu
shap~=0.47.2
boto3~=1.38.3
botocore~=1.38.3
textblob~=0.19.0
pymongo~=4.12.0
streamlit~=1.44.1
googletrans~=4.0.2
statsmodels~=0.14.4
loguru~=0.7.3
gym~=0.26.2
cryptography~=44.0.2
pyspark~=3.5.5